requirements.txt
streamlit
pandas
pillow
torch
torchvision
numpy